+++ src/FluxFlow.Components.Resources/IResourceDescriptorProvider.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Components.Resources.Contracts;
+
+namespace FluxFlow.Components.Resources;
+
+public interface IResourceDescriptorProvider
+{
+    IReadOnlyCollection<ResourceDescriptor> GetResources();
+}